api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
redis>=5.0.0
slowapi>=0.1.9
boto3>=1.34.0
python-dotenv>=1.0.0
structlog>=24.1.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
